% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/annoXeset.R
\name{annoXeset}
\alias{annoXeset}
\title{Function for cross annotating a mouse expression set with human attributes.}
\usage{
annoXeset(eset, AnnPkg.in)
}
\arguments{
\item{eset}{Mouse expression set}

\item{AnnPkg.in}{Annotation package that corresponds to the mouse expression set.}
}
\description{
This function takes a mouse expression set as input, and outputs a merged expression set with mouse and human annotations.
}
\examples{
eset.IL23.probes.Xanno.2 <- annoXeset(eset=eset_IL23,AnnPkg.in = "mouse4302.db")
}
\keyword{Xanno,}
\keyword{annotation,}
\keyword{cross}
\keyword{eset,}
\keyword{expression}
\keyword{human}
\keyword{mouse,}
\keyword{set,}

