% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/annoXMouseHuman.R
\name{annoXMouseHuman}
\alias{annoXMouseHuman}
\title{Function for annotating and cross annotating top table of mouse DEGs}
\usage{
annoXMouseHuman(top = top.A.flakVSflg, mouse = "mmusculus_gene_ensembl",
  human = "hsapiens_gene_ensembl", mouse.attr.in = c("ensembl_gene_id",
  "mgi_symbol"), human.attr.in = c("ensembl_gene_id", "hgnc_symbol",
  "wikigene_description"), top.tab = TRUE,
  AnnPkg = "mogene21sttranscriptcluster.db")
}
\arguments{
\item{top}{Top table as derived by calling the function toptable on an ebayes fit object}

\item{mouse}{The mouse mart to use in the biomaRt function}

\item{human}{The human mart to use in the biomaRt function}

\item{mouse.attr.in}{Mouse attributes to use from biomaRt.}

\item{human.attr.in}{Human attributes to use from biomaRt.}

\item{top.tab}{TRUE/FALSE if "top" is actually a top table}

\item{AnnPkg}{Mouse eset annotation package. So which package was used to annotate the mouse expression set.}
}
\description{
This function does cross annotation of a top table of mouse DEGs to human gene Symbols, making use of biomaRt
}
\examples{
annoXMouseHuman(top=top.A.flakVSflg)
}
\keyword{annotation}
\keyword{annotation,}
\keyword{cross}
\keyword{human,}
\keyword{mouse,}
\keyword{xannotation,}

