% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/auxBiplot.R
\name{auxBiplot}
\alias{auxBiplot}
\title{Plots a boxplot of qVar over different categories of BiVar. If model=TRUE, it overlays the estimated mean and SEM.}
\usage{
auxBiplot(dat, qVar, BiVar, qLab = qVar, BiLab = BiVar, model = TRUE,
  logbase = NULL, fdata = "", dir = getwd())
}
\arguments{
\item{dat:}{database}

\item{qVar:}{quantitative variable}

\item{BiVar:}{categorical variable}

\item{qLab:}{label for quantitative variable}

\item{BiLab:}{label for categorical variable}

\item{model=TRUE,}{overlays mean and SEM}

\item{logbase=NULL:}{if the data was transfromed using logb, this willl allows to plot results in the naturale scale, logbase is the base used in the log transformation}

\item{fdata=''}{Suffix Name for the plot; It would be composed as paste(dir,'/Boxplot',fdata,'_',qVar,'By',BiVar,sep='')}

\item{dir=getwd()}{directory to save plot and table with resutls}
}
\description{
It selects the N most dissimilar models among those in caret suite by maximizing the Jaccard dissimilarity between sets of models.
}
\examples{
auxBiplot(X,qVar='IL13',BiVar='Tissue',ylab=expression(log[2]('IL-13'),model=TRUE,dir=getwd(),logbase=NULL)
}

