% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/calcFCH.R
\name{calcFCH}
\alias{calcFCH}
\title{A function that given a matriz of log2FCH calculates the (signed) FCH}
\usage{
calcFCH(coefs.lg)
}
\arguments{
\item{coefs.lg:}{matrix of log fold changes, usually the estimated coeficients of the limma model (ebfit$coef)}
}
\description{
A function that reads GMT files and returns a list where each element is a list of genes
}
\examples{
A=rnorm(cbind(runif(10,-2,2),runif(10,-1,1)))
calcFCH(A)
}

