% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/CalculateCorrMatrix.R
\name{CalculateCorrMatrix}
\alias{CalculateCorrMatrix}
\title{Internal function of DoCorrelationPlotwithStarts.}
\usage{
CalculateCorrMatrix(db, method, abs.cor = FALSE)
}
\arguments{
\item{db:}{data base of continuos variables; columns are variables to be clustered.}

\item{method:}{correlation method (spearman, perason etc, default to spearman)}

\item{abs.cor:}{should the distance be 1-cor (FALSE,default) or 1-|cor| (TRUE). In teh later case, variables negative correlated will be as near as positive correlated}
}
\description{
It calculates the correlation matrix and pvalues and returns a list with components.
}
\examples{
# in this examples, I want to cluster samples
a<-CalculateCorrMatrix(exprs(reset)[1:10,])
print(round(a$cor,2))
print(1*(a$p<=0.05))
}

