% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/colChoice.2.R
\name{colChoice.2}
\alias{colChoice.2}
\title{Defining colors from a pattern}
\usage{
colChoice.2(colpat = "blue", colfacs)
}
\arguments{
\item{colpat}{is the color pattern that will guide choice of colors}

\item{colfacs}{is the factor that should be colored according to the color pattern}
}
\description{
a function that returns a vector of color from a given pattern
}
\examples{
colChoice.2.R(colplat="blue",colfacs = myfactor)
}

