% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/collapseduplicatesids.R
\name{collapseduplicatesids}
\alias{collapseduplicatesids}
\title{A function that collapse annotation of probes that mapp to more than one gene}
\usage{
collapseduplicatesids(gns)
}
\arguments{
\item{gns}{matrix of genes}
}
\description{
This function given a matrix of genes collapses annotation of probes that mapp to more than one gene
}

