% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/collapsefcn.R
\name{collapsefcn}
\alias{collapsefcn}
\title{Given a vector of characters a b c it returns a single character of the form a |b |c}
\usage{
collapsefcn(y = c("a", "b", "c"))
}
\arguments{
\item{y}{vector of characters}
}
\description{
Given a vector of characters a b c it returns a single character of the form a |b |c  . Internal function to use in annotation.
}
\examples{
collapsefcn()
}

