% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ComBat_MSF.R
\name{ComBat_MSF}
\alias{ComBat_MSF}
\title{Function that runs the original Combat script but returns different parts of the internal model instead of just the adjusted data. It's meant to be used when the design is complex and not fullly modeled,
 so one can get the estimated coeficients and play with it. I sed it in the IBM competition}
\usage{
ComBat_MSF(expression_xls, sample_info_file, type = "txt", write = T,
  covariates = "all", par.prior = T, filter = F, skip = 0,
  prior.plots = T)
}
\arguments{
\item{exprs:}{matrix of expression}

\item{g:}{factor defining the batches
#'}
}
\description{
Inputs and Outputs are the same as in ComBat.R
}
\examples{
abatch=ReadAffy()
abatch$Batch<-sapply(protocolData(abatch)$ScanDate,function(x){substr(x,1,10)}
eset<-rma(abatch)
out<-RunCombat_MSF(exprs(eset),eset$Batch)
}

