% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cor.prob.pearson.R
\name{cor.prob.pearson}
\alias{cor.prob.pearson}
\title{calculates the p-values for the pair-wise pearson correlation coefficient (called by CorrelationHeatmap)}
\usage{
cor.prob.pearson(x)
}
\arguments{
\item{x:}{matrix of observations by variables}
}
\description{
Given a matrix of observations by variables, calculates the pearson correlation between the variable(columns). columns must be named. It returns a simmetric matrix of number of variables, with p values for the correlation between row and column.
}
\examples{
cor.prob.pearson(db)
}

