% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cor.prob.spearman.R
\name{cor.prob.spearman}
\alias{cor.prob.spearman}
\title{calculates the p-values for the pair-wise spearman correlation coefficient (called by CorrelationHeatmap)}
\usage{
cor.prob.spearman(X, dfr = nrow(X) - 2)
}
\arguments{
\item{x:}{matrix of observations by variables}
}
\description{
Given a matrix of observations by variables, calculates the spearman correlation between the variable(columns). columns must be named. It returns a simmetric matrix of number of variables, with p values for the correlation between row and column.
}
\examples{
cor.prob.pearson(db)
}

