% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/CorrelationHeatmap_Samples.R
\name{CorrelationHeatmap_Samples}
\alias{CorrelationHeatmap_Samples}
\title{Function that given a data base of continuos variables, clusters it by variable and represents the associations as a heatmap-colored correlation matrix.}
\usage{
CorrelationHeatmap_Samples(db, method = "spearman", abs.cor = FALSE,
  hmeth = "average", max.stars = 0.05, breaks = seq(-1, 1, 0.25), ...)
}
\arguments{
\item{db:}{data base of continuos variables; columns are variables to be clustered.}

\item{method:}{correlation method (spearman, perason etc, default to spearman)}

\item{abs.cor:}{should the distance be 1-cor (FALSE,default) or 1-|cor| (TRUE). In teh later case, variables negative correlated will be as near as positive correlated}

\item{hmeth:}{aglomeration method to do the k-means clustering. default to average}

\item{max.stars:}{pvalue that will be represented by one asterisc (*) . If max.stars=0.05 (default), ***=0.001,**=0.01, *=0.05}

\item{breaks:}{color breaks for heatmap, defualt seq(-1,1,0.25);}

\item{print.corrcoef:}{Parameters to control the heatmap aestetics. Logicalin dicating if correlation coeficiets should be printed. Default to TRUE.}

\item{title:}{Parameters to control the heatmap aestetics. title for the plot}

\item{stars.size:}{Parameters to control the heatmap aestetics. size of the text overlay in the figure (correlations and p values)}
}
\description{
Function that given a data base of continuos variables, clusters it by variable and represents the associations as a heatmap-colored correlation matrix in the order established by undupervised clustering analysis. The correlation coeficient appears in the lower diagonal and the p-values  based * are in the upper diagonal.
}
\examples{
# in this examples, I want to cluster samples
CorrelationHeatmap_Samples(exprs(reset)[1:10,],print.corrcoef=F) #supress text
# in this examples, I want to cluster genes
CorrelationHeatmap_Samples(t(exprs(reset)[1:10,]), max.stars=0.1,print.corrcoef=F) #***=0.01,**=0.05, *=0.1
}

