% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/CreateOutputsMA.R
\name{CreateOutputsMA}
\alias{CreateOutputsMA}
\title{A function that creates a table with the results of limma models}
\usage{
CreateOutputsMA(ebfit, AnnPkg, Decision = NULL,
  coefs = c(1:ncol(ebfit$coef)), adjmeth = "fdr", mcut = 2, fname = "",
  pcut = 0.05, html.genetable = TRUE, exprset = NULL, annotTable = NULL,
  which.annot = c("ENTREZID", "SYMBOL", "GENENAME", "OMIM", "CHRLOC"),
  which.annot.unique = c("ENTREZID", "SYMBOL", "OMIM", "CHRLOC"),
  pPath = 0.1, pGO = pPath, nPath = 2, pPFAM = pPath)
}
\arguments{
\item{ebfit}{has estimates of contrasts}

\item{AnnPkg}{package used to annotate data (example: hgu133plus2.db)}

\item{Decision}{is the matrix with -1 0 1 indicating gene regulation}

\item{coefs}{is a matrix with the coefficients of the models}

\item{adjmeth}{is the criterion for adjusting pvalues}

\item{mcut}{is the cutoff for log2 fold-changes}

\item{fname}{is the output filename}

\item{pcut}{is the threshold for pvalues}

\item{html.genetable}{logical parameter to create html table with results}

\item{exprset}{is the matrix with gene expressions}

\item{annotTable}{is a table with annotation about the gens}

\item{which.annot}{is a selection of which annotations should be used}

\item{which.annot.unique}{is a selection of which annotations should be unique}

\item{pPath}{is the pvalue cutoff for pathway analysis}

\item{pGO}{is the pvalue cutoff for annotation in Gene Ontology Database}

\item{nPath}{?????}

\item{pPFAM}{is the cutoff for annotation in Protein Family DataBase}
}
\description{
a function that takes ebfit as input and create a table with DEG genes
}
\examples{
CreateOutputsMA(ebfit, AnnPkg, Decision=NULL, coefs=c(1:ncol(ebfit$coef)), adjmeth='fdr', mcut=2,
fname="", pcut=0.05, html.genetable =TRUE, exprset=NULL, annotTable=NULL,
which.annot= c("ENTREZID","SYMBOL","GENENAME","OMIM","CHRLOC"),
which.annot.unique=c("ENTREZID","SYMBOL","OMIM","CHRLOC"),
pPath=0.1, pGO=pPath, nPath=2, pPFAM=pPath)
}

