% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/deletePathway.R
\name{deletePathway}
\alias{deletePathway}
\title{A function that deletes pathways whose frequency is greater than 5}
\usage{
deletePathway(SET, MyGroups = mypathways, max = 9999, min = 0)
}
\arguments{
\item{SET}{a data frame}

\item{MyGroups}{is the data frame column that contains the names of pathways}

\item{max}{is the max acceptable frequency}

\item{min}{is the minimum acceptable frequency}
}
\description{
a function that examines a frequency table and conserves the elements
with frequency greater than min and less than max parameters
}
\examples{
deletePathway(SET,MyGroups=mypathways,max=9999,min=0)
}

