% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dis.cors.R
\name{dis.cors}
\alias{dis.cors}
\title{Internal function of DoCorrelationPlotwithStarts.}
\usage{
dis.cors(x, abs = FALSE, method = "spearman", ...)
}
\arguments{
\item{x:}{symetrix matrix indicating the  correlation coefficient for each column/row pair of variables}

\item{method:}{correlation method (spearman, perason etc, default to spearman)}

\item{abs:}{should the distance be 1-cor (FALSE,default) or 1-|cor| (TRUE). In teh later case, variables negative correlated will be as near as positive correlated}
}
\description{
Distance matrix based on correlations.
}
\examples{
# in this examples, I want to cluster samples
a<-dis.cors(cor(exprs(reset)[1:10,]), abs=TRUE,method='pearson')
}

