% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/doBarPlot.R
\name{doBarPlot}
\alias{doBarPlot}
\title{A function that creates a barplot with a table for contrasts}
\usage{
doBarPlot(d, nn, ID, Group)
}
\arguments{
\item{d}{data frame with expression data.}

\item{nn}{name of the gene}

\item{ID}{name of the variable with subject ID}

\item{Group}{name of factor with groups to be compared}
}
\description{
This function creates a barplot with a table with estimated contrasts.
}
\examples{
doBarPlot(d,nn,ID,Group)
}
\keyword{,}
\keyword{barplot}
\keyword{expression}
\keyword{vizualisation}

