% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/DoCluster.R
\name{DoCluster}
\alias{DoCluster}
\title{Save output of a heatmap built from a decision matrix and coefficients from limma model}
\usage{
DoCluster(Decision, ebfit, reset, dmeth = "euclidean", hmeth = "complete",
  cols = NULL)
}
\arguments{
\item{Decision}{is a matrix indicating up(1), down(-1) regulation for each contrast is a gene list.}

\item{ebfit}{is a matrix of coefficients for each contrast}

\item{reset}{is the expression set used as input for the limma model}

\item{dmeth}{is the method for distance}

\item{hmeth}{is the agglomerative method used for the hierarchical clustering}

\item{cols}{is a vector defining how the columns in the hierarchical heatmap should be ordered}

\item{Group}{is a vector indicating the levels associated with columns in the expression sets}

\item{Patients}{is a vector indicating the patients IDs associated with columns in the expression sets}
}
\description{
a function that preprocess output from limma model and draws a heatmap according to a set of parameters
}
\examples{
DoCluster(Decision = D, ebfit = fit, reset = eset, dmeth = "euclidean" , hmeth = "complete" , cols =NULL,Group=Group, Patients = IDs)
}

