% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/doClusterTableFCH_fromMatrix.R
\name{doClusterTableFCH_fromMatrix}
\alias{doClusterTableFCH_fromMatrix}
\title{A function that creates a heatmap with column indicators on top of each column.}
\usage{
doClusterTableFCH_fromMatrix(mat, ps = rownames(mat), colfac,
  facsepr = NULL, hmeth = "average", dmeth = cor.dist, coefs, fdrs,
  main = "", ColD = FALSE, ss = ps, gn = NULL, breaks = NULL,
  cexg = 1, margins = c(5, 20))
}
\arguments{
\item{mat}{numeric matrix to cluster. Usually th raw expression data.}

\item{facsepr}{factor to separate columns in the heatmaps}

\item{hmeth}{aglomeration strategy}

\item{dmeth}{aglomeration distance to cluster}

\item{coefs}{estimates to be drawn tin the tabl (eg lgFCHs)}

\item{fdrs}{p values or fdr indicating significance of the estimates; should be the same dim as coefs}

\item{ColD}{logical TRUE if cluster by rows and columns}

\item{gn}{gene name}

\item{colfc}{factor to add colored legend to the heatmap}

\item{sn}{edited names for the rows (if different than the rownames) eg symbols for probesets}
}
\description{
This function creates a nice heatmap for publication. It puts on top of each column a color indicator of what sample type it is, does clustering, and adds a table in the right margin of the figure with logFoldChanges and respective significance indicators (stars).
}
\examples{
doClusterTableFCH_fromMatrix(mat,ps=rownames(mat),colfac,facsepr=NULL,hmeth='average',dmeth=cor.dist,coefs,fdrs,main="", ColD=FALSE, ss=ps, gn=NULL, breaks=NULL,cexg=1,margins=c(5,20) )
}
\keyword{,}
\keyword{expression}
\keyword{heatmap}
\keyword{vizualisation}

