% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/DoCorrelationPlotwithStars.R
\name{DoCorrelationPlotwithStarts}
\alias{DoCorrelationPlotwithStarts}
\title{Function that given a correlation matrix and the pvalues, makes a heatmap representing the pairwise correlation.}
\usage{
DoCorrelationPlotwithStarts(db.cor, db.pcor, max.stars = 0.05,
  print.corrcoef = TRUE, title = "Correlations", breaks = seq(-1, 1,
  0.25), stars.size = 2, ...)
}
\arguments{
\item{db.cor:}{symetrix matrix indicating the  correlation coefficient for each column/row pair of variables}

\item{db.pcor:}{symetrix matrix indicating the p-value associated with the correlation coefficient for each column/row pair of variables}

\item{max.stars:}{pvalue that will be represented by one asterisc (*) . If max.stars=0.05 (default), ***=0.001,**=0.01, *=0.05}

\item{print.corrcoef:}{Logicalindicating if correlation coeficiets should be printed. Default to TRUE.}

\item{title:}{title for the plot}

\item{breaks:}{color breaks for heatmap, defualt seq(-1,1,0.25);}

\item{stars.size:}{size of the text overlay in the figure (correlations and p values)}
}
\description{
Function that given a correlation matrix and the pvalues, makes a heatmap representing the pairwise correlation. It's used by CorrelationHeatmap_Samples functions. The correlation coeficient appears in the lower diagonal and the p-values  based * are in the upper diagonal.
}
\examples{
DoCorrelationPlotwithStarts(db.cor, db.pcor, max.stars=0.05, print.corrcoef=TRUE)
}

