% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/DoHeatmap.R
\name{DoHeatmap}
\alias{DoHeatmap}
\title{A function that draws a heatmap based on a matrix}
\usage{
DoHeatmap(mat, facsepr = NULL, colfac, symb = NULL, dmeth = cor.dist,
  hmeth = "average", cex.genes = 0.5, margins = c(5, 15), ColD = FALSE,
  main = "", scl = "row", breaks = NULL, cexCol = 1,
  colscheme = "gray", ...)
}
\arguments{
\item{mat}{a matrix with genes in the rows and samples in the columns}

\item{facsepr}{a factor in the phenotype data that separates columns in the heatmap}

\item{colfac}{a factor in the phenotype data that will add colors to the top according to different levels}

\item{symb}{a character vector that indicates gene symbols for each row}

\item{dmeth}{the method to measure distance between matrices rows}

\item{hmeth}{the method for the hierarchical clustering}

\item{cex.genes}{is the proportion to which genes symbols should be expanded in the graphic}

\item{margins}{is the definition of vertical and horizontal margins to plot the heatmap}

\item{ColD}{is a logical parameter that indicates if the columns should be clustered}

\item{main}{is the title for the heatmap}

\item{scl}{is an indicator if data should be scaled by row, column or both}

\item{breaks}{enable color transition at specified limits}

\item{cexCol}{is the proportion to which columns labels should be expanded in the graphic}

\item{colscheme}{is the schema for colors in the body of the heatmap}
}
\description{
a function that is supposed to draw a heatmap based on a matrix of gene expressions
}
\examples{
DoHeatmap(mat = exprs(eset),facsepr = mygroup1, colfac = mygroup2, symb = genesymbols, dmeth=cor.dist, hmeth='average',
cex.genes=0.5, margins = c(5,15), main='', scl = row, breaks = NULL, cexCol = 1, colscheme = 'gray')
}

