% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/DoHeatmapSamples.R
\name{DoHeatmapSamples}
\alias{DoHeatmapSamples}
\title{A function that draws a heatmap based on a matrix}
\usage{
DoHeatmapSamples(exprs, corm = "pearson", colfac = NULL, dmeth = euc,
  hmeth = "average", cex.genes = 0.5, margins = c(5, 15), main = "",
  scl = "none", thebreaks = NULL, samplelabels = NULL, colcol = NULL,
  ...)
}
\arguments{
\item{exprs}{is a matrix of expressions}

\item{corm}{is the metric for correlation evaluation}

\item{colfac}{a factor in the phenotype data that will add colors to the top according to different levels}

\item{dmeth}{the method to measure distance between matrices rows}

\item{hmeth}{the method for the hierarchical clustering}

\item{cex.genes}{is the proportion to which genes symbols should be expanded in the graphic}

\item{margins}{is the definition of vertical and horizontal margins to plot the heatmap}

\item{main}{is the title for the heatmap}

\item{scl}{is an indicator if data should be scaled by row, column or both}

\item{thebreaks}{enable color transition at specified limits}

\item{samplelabels}{keeps a vector of characters to be used in the heatmap columns}

\item{colcol}{is the color schema for the rows}

\item{ColD}{is a logical parameter that indicates if the columns should be clustered}

\item{colscheme}{is the schema for colors in the body of the heatmap}
}
\description{
a function that is supposed to draw a heatmap based on a matrix of gene expressions
}
\examples{
DoHeatmap(mat = exprs(eset),facsepr = mygroup1, colfac = mygroup2, symb = genesymbols, dmeth=cor.dist, hmeth='average',
cex.genes=0.5, margins = c(5,15), main='', scl = row, breaks = NULL, cexCol = 1, colscheme = 'gray')
}

