% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/doVenn.R
\name{doVenn}
\alias{doVenn}
\title{A function to draw Venn Diagrams from Decision Matrices}
\usage{
doVenn(theD, ps, fname, shape = "circles", circle.names = NULL,
  show = list(Faces = FALSE, DarkMatter = FALSE))
}
\arguments{
\item{theD}{is the decision matrix that indicates if genes are differentially expressed}

\item{ps}{is the vector with identification for the probesets}

\item{fname}{is the prefix for the file to save the output}

\item{shape}{is the parameter that defines the shape for the Venn diagram figure}

\item{circle.names}{is a vector with the identifiers for the circles}

\item{show}{is a list if some parameters for the Vennerable Venn Diagram}
}
\description{
a function that is supposed to draw a heatmap based on a matrix of gene expressions
}
\examples{
doVenn(theD = D, ps = probeset, fname = 'filename', shape='circles', circle.names=NULL, show = list(Faces=FALSE,DarkMatter=FALSE))
}

