% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dowordcloud.R
\name{dowordcloud}
\alias{dowordcloud}
\title{Creates a wordcloud representin Fold changes and color coded to significance}
\usage{
dowordcloud(vars, Dmat, title, addFCH = NULL, logEst = FALSE,
  leg_wcloud = c("Up-(p>0.1)", "Up-(p<0.1)", "Up-(p<0.05)", "Down-(p>0.1)",
  "Down-(p<0.1)", "Down-(p<0.05)"), col_wcloud = c("lightpink2", "red",
  "darkred", "lightblue", "deepskyblue2", "blue3"))
}
\arguments{
\item{vars}{variables to add to wordcloud}

\item{Dmat}{matrix where rows are features to appear as words and columns are elements of the comparison to be used. column Estimate and pvalue will be used for size/colors of words}

\item{addFCH}{it contains numbers that will be ploted as to give an idea of which fold changes represents teh size.}

\item{logEst}{should be False unless to want to size it by log2 scale}

\item{leg_wcloud}{legend for colors}

\item{col_wcloud}{colors}
}
\description{
the features in vars will have size proportional to the column Estimate in Dmat and color coded to column pvalues. Positive estimates (up-regulated) will be in shades of red depending on significance, and negative estimates (down-regulated) will be in shades of blue)
}
\examples{
dowordcloud()
}
\keyword{;}
\keyword{plotting}
\keyword{wordcloud}

