% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ellipsoid.Fox.R
\name{ellipsoid.Fox}
\alias{ellipsoid.Fox}
\title{Function from John Fox that is called by scatter3d.Fox.}
\usage{
ellipsoid.Fox(center = c(0, 0, 0), radius = 1, shape = diag(3),
  segments = 51)
}
\arguments{
\item{center}{center of the ellipsoid}

\item{radius}{radius of the ellipsoid}

\item{shape}{shape of the ellipsoid}
}
\description{
It returns a matriz with 3 columns  to be used to draw confidence ellipsoid in 3d.
}
\examples{
ellipsoid.Fox()
}

