% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/firstelementfcn.R
\name{firstelementfcn}
\alias{firstelementfcn}
\title{It returns the first element non-NA from a vector.}
\usage{
firstelementfcn(y)
}
\arguments{
\item{y}{vector of characters}
}
\description{
Internal function to use in annotation.
}
\examples{
firstelementfcn(LETTERS[1:10])
}

