% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/form1list2D.R
\name{from1list2D}
\alias{from1list2D}
\title{A function that converts a list of genelist into a Decision matrix}
\usage{
from1list2D(list)
}
\arguments{
\item{list:}{list of L vectors containing gene lists}
}
\description{
Converts a list of L vectors containing features into a binary indicator matrix of nxL where L is the length of the universe of genes in the List
}
\examples{
L<-readGmt('KruegerSYMBOL2015.gmt')
D<-from1list2D(L)
}

