% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/FromEbfit2DEGTable.R
\name{FromEbfit2DEGTable}
\alias{FromEbfit2DEGTable}
\title{creates an output table including genes that are declared differentially expresed (given the user-speceified cut=offs) in at least one of the comparisons embeded in ebfit}
\usage{
FromEbfit2DEGTable(ebfit, adj = "BH", mcut = 2, pcut = 0.05,
  annot = TRUE, annotTab = NULL)
}
\arguments{
\item{ebfit}{output of eBayes}

\item{adj}{adjustment method for multiple hypothesis}

\item{mcut}{cutoff for FCH (to be used for Status columns)}

\item{pcut}{cutoff for pvalue/FDR  (to be used for Status columns)}

\item{annot}{should the dat be annotated? If TRUE it will check if tehre is annotTab or otherwise whatever annotation is stored in ebfit$genes}

\item{annotTab}{annotation database}
}
\description{
For each contratst in ebfit, a matrix with lgFCH, FCH,pvalue , fdr and Status (1 up,-1 doen,0 non sign) is created using pre-specified cut-oofs
}
\examples{
TabDEG<-FromEbfit2DEGTable(ebfit,pcut=0.05,mcut=2)
}

