% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/FromEbfit2Table.R
\name{FromEbfit2Table}
\alias{FromEbfit2Table}
\title{creates an output table including all genes in ebfit}
\usage{
FromEbfit2Table(ebfit, adj = "BH", mcut = 2, pcut = 0.05, annot = FALSE,
  annotTab = NULL)
}
\arguments{
\item{ebfit}{output of eBayes}

\item{adj}{adjustment method for multiple hypothesis}

\item{mcut}{cutoff for FCH (to be used for Status columns)}

\item{pcut}{cutoff for pvalue/FDR  (to be used for Status columns)}

\item{annot}{should the dat be annotated? If TRUE it will check if tehre is annotTab or otherwise whatever annotation is stored in ebfit$genes}

\item{annotTab}{annotation database}
}
\description{
For each contratst in ebfit, a matrix with lgFCH, FCH,pvalue , fdr and Status (1 up,-1 doen,0 non sign) is created using pre-specified cut-oofs
}
\examples{
Tab<-FromEbfit2Table(ebfit,pcut=0.05,mcut=2)
}

