% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fromPtoCol.R
\name{fromPtoCol}
\alias{fromPtoCol}
\title{Internal function of dowordcloud.}
\usage{
fromPtoCol(p, fch = NULL, col_wcloud = c("mistyrose3", "red3", "darkred",
  "greenyellow", "green", "darkgreen"))
}
\arguments{
\item{p}{p values determining colors}

\item{fch}{estimates (fold changes) red/blue colors will be used for positive/negatives values}

\item{col_wcloud}{colors}

\item{leg_wcloud}{legend for colors}
}
\description{
Transforms pvalues to colors
}

