% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getAnnotationTable.R, R/getAnnotationTable_JCR.R
\name{getAnnotationTable}
\alias{getAnnotationTable}
\title{A function that creates an Annotation table for microorrays experiments.}
\usage{
getAnnotationTable(IDS, AnnPkg, w.annot = c("ENTREZID", "SYMBOL", "GENENAME",
  "OMIM", "CHRLOC", "PATH", "GO"), w.annot.unique = c("ENTREZID", "SYMBOL",
  "OMIM", "CHRLOC"))

getAnnotationTable(IDS, AnnPkg, w.annot = c("ENTREZID", "SYMBOL", "GENENAME",
  "OMIM", "CHRLOC", "PATH", "GO"), w.annot.unique = c("ENTREZID", "SYMBOL",
  "OMIM", "CHRLOC"))
}
\arguments{
\item{IDS}{identification for the probesets}

\item{AnnPkg}{annotation package (e.g. hgu133plus2.db)}

\item{IDS:}{type of methods. Options are Regression, Classification adn Dual use}

\item{AnnPkg:}{named of the Array annotation data (ie, hgu133plus2.db)}

\item{w.annot:}{name of the information to be included in the table}

\item{w.annot.unique:}{which columns should be used to solved duplicated annotation; ie when a probe have two gene names.}

\item{w.annot:}{which information should be annotated}

\item{w.annot.unique:}{unique annotation to be done}
}
\description{
A function that creates an Annotation table for microorrays experiments. It must be associated with an existing Array annotation data (ie hgu133plus2.db) where annotations has been
assembled using data from public repositories

this function annotates probeids by using AnnotationDbi package
}
\examples{
library(hgu133plus2.db)
AnnPkg <- hgu133plus2.db
getAnnotationTable(IDS=featureNames(eset), AnnPkg, w.annot=c("ENTREZID","SYMBOL","GENENAME"),w.annot.unique=c("ENTREZID"))
getAnnotationTable<-function(IDS, hgu133plus2.db, w.annot=c("ENTREZID","SYMBOL","GENENAME","OMIM","CHRLOC","PATH","GO"),w.annot.unique=c("ENTREZID","SYMBOL","OMIM","CHRLOC"))
}
\keyword{annotation,}
\keyword{gene}
\keyword{ontology,}
\keyword{probeids}

