% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getgray.R
\name{getgray}
\alias{getgray}
\title{A function that assigns levels of gray to a vector}
\usage{
getgray(g)
}
\arguments{
\item{g}{vector}
}
\description{
This function creates a vector with differents levels of gray in concordance with the levels of a given vector
}
\examples{
getgray(c("mild","moderate","severe"))
}

