% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getMeanCI.by.Path.R
\name{getMeanCI.by.Path}
\alias{getMeanCI.by.Path}
\title{Function given a value in y and a collection of pathways}
\usage{
getMeanCI.by.Path(SETs, y)
}
\arguments{
\item{SETs}{matrix of number_of_genes by number_of_pathways. TRUE of FALSE if the gene belong to the pathway}

\item{y}{named vector of size number_of_genes.}
}
\description{
Given a collection of pathwyas in SETs and a value of an attribute in y, it summarizes the mean, CI and p values foe y for each patway
}

