% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getModels_Caret.R, R/getModels_Caret_JCR.R
\name{getModels_Caret}
\alias{getModels_Caret}
\title{A function that tells you what are most N disimilar models available in caret for a certain problem (tag)}
\usage{
getModels_Caret(method = "Regression", N = 5, include = "")

getModels_Caret(method = "Regression", N = 5, include = "")
}
\arguments{
\item{method}{is the kind of model Classification / Regression}

\item{N}{the number of dissimilar models to be extracted}

\item{include}{a character to separate the names of the columns}

\item{method:}{type of methods. Options are Regression, Classification adn Dual use}

\item{N:}{number of methods to select}

\item{include:}{name of a method to be included in the set}
}
\description{
It selects the N most dissimilar models among those in caret suite by maximizing the Jaccard dissimilarity between sets of models.

this function extracts a list of models in caret that are dissimilar
}
\examples{
getModels_Caret<-function(method="Regression",N=5,include='cubist')
getModels_Caret('Regression',5,'')
}
\keyword{,}
\keyword{caret}
\keyword{classification}
\keyword{regression}

