% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getMultipleMeanandCI.by.Path.R
\name{getMultipleMeanCI.by.Path}
\alias{getMultipleMeanCI.by.Path}
\title{Function given a collection of pathways, returns the summary ( score for)mean, CI, p value etc) for multiple comparisons at once}
\usage{
getMultipleMeanCI.by.Path(STATs, SETs, fname)
}
\arguments{
\item{SETs}{matrix of number_of_genes by number_of_pathways. TRUE of FALSE if the gene belong to the pathway}

\item{fname}{name of the file to save results}

\item{STATS}{matrix of number_of_genes by number_of_comparisons with the estimates}
}
\description{
Given a collection of pathwyas in SETs and the value of an estimate (ie lg FCH) in each column of STATs, it returns the summary
}
\examples{
z=getMultiplMeanCI.by.Path(ebfit$coef,SETs,'MeanFCH4Comparisons')
}

