% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getMultipleZscore.by.Path.R
\name{getMultipleZscore.by.Path}
\alias{getMultipleZscore.by.Path}
\title{Function given a collection of pathways, returns the Z score for multiple comparisons at once}
\usage{
getMultipleZscore.by.Path(STATs, SETs, fname)
}
\arguments{
\item{SETs}{matrix of number_of_genes by number_of_pathways. TRUE of FALSE if the gene belong to the pathway}

\item{fname}{name of the file to save results}

\item{STATS}{matrix of number_of_genes by number_of_comparisons with the T statistics}
}
\description{
Given a collection of pathwyas in SETs and the value of a T statistic in each column of T, it returns the Z score in a matrix of number_of_pathwyas by number_of_comparisons
}
\examples{
z=getMultipleZscore.by.Path(ebfit$t,SETs,'Zscore4Comparisons')
}

