% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getPCAs.R
\name{getPCAs}
\alias{getPCAs}
\title{Extraction of principal components}
\usage{
getPCAs(eset, maxpc = 3)
}
\arguments{
\item{eset}{is an expression set}

\item{maxpc}{is the maximum number of principal components to be extracted}
}
\description{
a function that extracts principal components from an expression set and output the scores of most important components and variance explained
the original function would output all results in a data frame. This function outputs to a list
}
\examples{
getPCAs(eset,maxpc=3)
}

