% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getstar.R
\name{getstar}
\alias{getstar}
\title{A function that returns an object of type character equal to 1, 2, 3 or 4 stars according to the value of a given object of type numeric}
\usage{
getstar(p, pmax = 0.05, parenthesis = FALSE)
}
\arguments{
\item{p}{numeric object. Usually a p value of a Test.}

\item{pmax}{numeric object. Usually the maximun permisible Type I Error of a Test.}

\item{parenthesis}{logical TRUE if parenthesis are required to wrap the stars}
}
\description{
This function creates an object of type character equal to 1, 2, 3 or 4 stars (*) for diferents ranges of p values. To smaller p values correspond more number of stars.
}

