% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getSymbolByX.R
\name{getSymbolByX}
\alias{getSymbolByX}
\title{A function that}
\usage{
getSymbolByX(Pmat, X, descend = TRUE)
}
\arguments{
\item{Pmat}{numeric matrix}

\item{X}{expression set}

\item{descend}{logical object. The default is TRUE. If TRUE the output will be sort on decreassing order.}
}
\description{
This function
}

