% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getZs.R
\name{getZs}
\alias{getZs}
\title{A function that calculates Z-scores and p-values}
\usage{
getZs(SETs, T)
}
\arguments{
\item{SETs}{numeric matrix}
}
\description{
This function given a set of values, returns Z-scores, p-values and p-values adjusted
}

