% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getZscore.by.Path.R
\name{getZscore.by.Path}
\alias{getZscore.by.Path}
\title{Function given a T statistic and a collection of pathways, returns the Z score}
\usage{
getZscore.by.Path(SETs, T)
}
\arguments{
\item{SETs}{matrix of number_of_genes by number_of_pathways. TRUE of FALSE if the gene belong to the pathway}

\item{y}{named vector of size number_of_genes.}
}
\description{
Given a collection of pathwyas in SETs and the value of a T statistic in T, it returns the Z score for each patway
}
\examples{
z=getZscore.by.Path(SETs,ebfit$t[,1])
}

