% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ggDendogram_MSF.R
\name{ggDendrogram_MSF}
\alias{ggDendrogram_MSF}
\title{Function that prints a dendogram using ggplot. Was modified from ggdendrogram}
\usage{
ggDendrogram_MSF(data, segments = TRUE, labels = TRUE, leaf_labels = TRUE,
  rotate = FALSE, theme_dendro = TRUE, size = 12, ...)
}
\arguments{
\item{segments}{If TRUE, show line segments}

\item{labels}{if TRUE, shows segment labels}

\item{leaf_labels}{if TRUE, shows leaf labels}

\item{rotate}{if TRUE, rotates plot by 90 degrees}

\item{theme_dendro}{if TRUE, applies a blank theme to plot (see theme_dendro)}

\item{...}{other parameters passed to geom_text}

\item{Either}{a dendro object or an object that can be coerced to class dendro using the dendro_data function, i.e. objects of class dendrogram, hclust or tree}
}
\description{
Function that prints a dendogram using ggplot. Was modified from ggdendrogram . Refer to ggdendro package http://www.icesi.edu.co/CRAN/web/packages/ggdendro/vignettes/ggdendro.pdf . Full description of how to print dendograms here: http://rstudio-pubs-static.s3.amazonaws.com/1876_df0bf890dd54461f98719b461d987c3d.htmlThe correlation coeficient appears in the lower diagonal and the p-values  based * are in the upper diagonal.
}
\examples{
ggDendrogram_MSF(DATA, rotate=FALSE)
}

