% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/givecolnames.R
\name{givecolnames}
\alias{givecolnames}
\title{A function that changes the column names of a matrix by adding a suffix function unique does keep NA in the output}
\usage{
givecolnames(mat, suffix)
}
\arguments{
\item{mat:}{matrix}

\item{suffix:}{character}
}
\description{
A function that changes the column names of a matrix by adding a suffix function unique does keep NA in the output
}
\examples{
givecolnames(ebfit$coef,'lgFCH')
givecolnames(ebfit$p,'p')
}

