% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/makeGSEApheno.R
\name{makeGSEApheno}
\alias{makeGSEApheno}
\title{A function that creates genotypes for GSEA}
\usage{
makeGSEApheno(G)
}
\arguments{
\item{G}{a vector with the identification of the genotypes}
}
\description{
a function that creates genotypes for GSEA
}
\examples{
makeGSEApheno(G=D)
}

