% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/makenames_MSF.R
\name{makenames_MSF}
\alias{makenames_MSF}
\title{A function that given common gene names in dermatology, create expression with special character (like greek symbols).}
\usage{
makenames_MSF(vs)
}
\arguments{
\item{vs:}{a gene name}
}
\description{
very useful for title and axis in plots.
}
\examples{
makenames_MSF('IL17')
}

