% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{A function that replace values in a vector}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{vector of values}

\item{replacement}{vector of values}

\item{x}{vector of values}
}
\description{
This function replaces in a vector the values that satisfice a given pattern by other values also given
}
\examples{
mgsub(c(0,1),c("No","Yes"),x=c(1,1,1,0,1,0,0,1,0,1))
}

