% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mylog10.R
\name{mylog10}
\alias{mylog10}
\title{A function that returns the base 10 logarithm of the components of a numeric vector}
\usage{
mylog10(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function given a numeric vector calculates the base 10 logarithm of the components
}

