% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mylog2.R
\name{mylog2}
\alias{mylog2}
\title{A function that returns the base 2 logarithm of the components of a numeric vector}
\usage{
mylog2(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function given a numeric vector calculates the base 2 logarithm of the components
}

