% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/nmissing.R
\name{nmissing}
\alias{nmissing}
\title{Returns the number of missing values in a vector}
\usage{
nmissing(x)
}
\arguments{
\item{x}{vector}
}
\description{
Given a vector the function calculates the number of missing values
}

