% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prepareGSEArnk.R
\name{prepareGSEArnk}
\alias{prepareGSEArnk}
\title{A function that prepares GSEA ranked data}
\usage{
prepareGSEArnk(coefs, Annpkg = NULL, fname, symbs = NULL)
}
\arguments{
\item{Annpkg}{is the package with annotation for the platform that is being used}

\item{fname}{is the prefix for the files to be output}

\item{symbs}{is a vector with genes symbols}

\item{coef}{a matrix with coefficients of a linear model}
}
\description{
a function that prepares preRanked Gsea files for different contrasts
}
\examples{
prepareGSEArnk(coefs = coefs, Annpkg = hgu133plus2.db, fname='mycontrasts',symbs=NULL)
}

