% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prettyp.R
\name{prettyp}
\alias{prettyp}
\title{A function that express a number using scientific notation or an specific number of decimals}
\usage{
prettyp(p, dig = 2)
}
\arguments{
\item{p}{Numeric object}
}
\description{
This function express a number using scientific notation if it is smaller than a given power of 10 or rounded to a given number of decimals
}

