% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readGmt.R
\name{readGmt}
\alias{readGmt}
\title{A function that reads GMT files}
\usage{
readGmt(fname)
}
\arguments{
\item{fname:}{filename}
}
\description{
A function that reads GMT files and returns a list where each element is a list of genes
}
\examples{
readGmt('KruegerSYMBOL2015.gmt')
}

