% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readGmt2.R
\name{readGmt2}
\alias{readGmt2}
\title{Function for reading a GMT file of gene set collection}
\usage{
readGmt2(fname)
}
\arguments{
\item{fname}{File name of the .gmt file. The format of the file should be: 1.Column=Geneset Name ; 2.Column=Geneset Description ; 3.-N.Column=GENE SYMBOLS}
}
\description{
This function reads a gene set collection GMT file, and stores the output in a list of genesets, geneset.names, and  geneset.descriptions. This version is adapted from the GSA package by Robert Tibshirani, with some modifications with respect to the possible empty cells, making it more efficient.
}
\examples{
jkr.list <- readGmt(fname = "./KruegerSYMBOL.gmt")
}
\keyword{gene}
\keyword{gmt,}
\keyword{sets}

