% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/renamecols.R
\name{renamecols}
\alias{renamecols}
\title{A function that changes the names of a dataframe columns}
\usage{
renamecols(db, suffix)
}
\arguments{
\item{db}{dataframe}

\item{suffixb}{dataframe}
}
\description{
This function changes the names of a dataframe columns adding a suffix to the current names (uses "_" to separate coulmn name and suffix)
}

