% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Resume.R
\name{Resume}
\alias{Resume}
\title{A function that prepare the results of a linear model fit}
\usage{
Resume(ebfit, AnnPkg, Decision = NULL, coefs = c(1:ncol(ebfit$coef)),
  adjmeth = "fdr", method = "separate", mcut = 2, fname = "",
  pvalue = 0.05, pPath = 0.05, pGO = 0.05, nPath = 2, ann_ind = TRUE,
  exprset = NULL, which.anncol = c(1:3, 6:7, 9, 11, 12))
}
\arguments{
\item{ebfit}{output of eBayes}

\item{Decision}{output of decideTests}

\item{coefs}{vector of character strings that contains the coeficients to be included}

\item{adjmeth}{character string specifying p-value adjustment method.
Possible values are "none", "BH", "fdr" (equivalent to "BH"), "BY" and "holm".}

\item{method}{character string specify how probes and contrasts are to be combined in the multiple testing strategy.
Choices are "separate", "global", "hierarchical", "nestedF" or any partial string.}

\item{mcut}{numeric value, folchange of cut}

\item{fname}{}

\item{pvalue}{critical pvalue}

\item{pPath}{numeric value, ussually number of pathways}

\item{pGO}{numeric value}

\item{nPath}{number of Pathways}

\item{ann_ind}{logical value equal TRUE by default}

\item{exprset}{expression set}

\item{which.anncol}{numeric vector with annotation columns numbers}

\item{ANNPkg}{annotation package}
}
\description{
This function make a Resume of a microarray linear model fit results after eBayes function
}

