% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RunCombat.R
\name{RunCombat}
\alias{RunCombat}
\title{Function that runs the original Combat script where no model matrix of covariates is needed}
\usage{
RunCombat(eset, g, covariates = NULL)
}
\arguments{
\item{eset:}{object of class ExpressionSet}

\item{g:}{factor defining the batches}

\item{covariates:}{names of the covariates to be found in pData(eset)}
}
\description{
It accpets an ExpressionSet and returns also an ExpressionSet. If you have a matrix, use RunCombatMat
}
\examples{
abatch=ReadAffy()
abatch$Batch<-sapply(protocolData(abatch)$ScanDate,function(x){substr(x,1,10)}
eset<-rma(abatch)
RunCombat(eset,eset$Batch)
}

