% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RunCombat_MSF.R
\name{RunCombat_MSF}
\alias{RunCombat_MSF}
\title{Function that runs the original Combat script where no model matrix of covariates is needed. It calls a version created by MSF to return several parts of the internal modeling}
\usage{
RunCombat_MSF(eset, g, covariates = NULL)
}
\arguments{
\item{eset:}{object of class ExpressionSet}

\item{g:}{factor defining the batches}

\item{covariates:}{names of the covariates to be found in pData(eset)}
}
\description{
See ComBat_MSF. It accepts an ExpressionSet and returns also an ExpressionSet.
}
\examples{
abatch=ReadAffy()
abatch$Batch<-sapply(protocolData(abatch)$ScanDate,function(x){substr(x,1,10)}
eset<-rma(abatch)
RunCombat_MSF(eset,eset$Batch)
}

