% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RunCombatMat.R
\name{RunCombatMat}
\alias{RunCombatMat}
\title{Function that runs the original Combat script where no model matrix of covariates is needed.}
\usage{
RunCombatMat(eset, g)
}
\arguments{
\item{exprs:}{matrix of expression}

\item{g:}{factor defining the batches
#'}
}
\description{
It accepts an a matrix and gives back a matrix
}
\examples{
abatch=ReadAffy()
abatch$Batch<-sapply(protocolData(abatch)$ScanDate,function(x){substr(x,1,10)}
eset<-rma(abatch)
RunCombatMat(exprs(eset),eset$Batch)
}

