% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sciNotation.R
\name{sciNotation}
\alias{sciNotation}
\title{A function that returns a scientific notation of a p value}
\usage{
sciNotation(x, digits = 2, maxexp = 16)
}
\arguments{
\item{x}{Numeric object. Number to be expressed using scientific notation.}

\item{digits}{Numeric object of type integer. Number of digits.}

\item{maxexplogical}{Numeric object of type integer. Maximun exponential admissible.}
}
\description{
This function creates an expression of a p value using scientific notation
}
\examples{
sciNotation(x=0.00000000000068,digits=2,maxexp=16)
}

