% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/stackchar.R
\name{stackchar}
\alias{stackchar}
\title{A function that concatenates a vector of character using sperator sep}
\usage{
stackchar(X, sep)
}
\arguments{
\item{x:}{vector of characters}

\item{sep:}{separator to be used}
}
\description{
A function that concatenates a vector of character using sperator sep
}
\examples{
stackchar(LETTERS[1:10],",")
}

