% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/summarySEwithin.R
\name{summarySEwithin}
\alias{summarySEwithin}
\title{## Summarizes data, handling within-subjects variables by removing inter-subject variability.}
\usage{
summarySEwithin(data = NULL, measurevar, betweenvars = NULL,
  withinvars = NULL, idvar = NULL, na.rm = FALSE, conf.interval = 0.95,
  .drop = TRUE)
}
\arguments{
\item{data:}{a data frame.}

\item{measurevar:}{the name of a column that contains the variable to be summariezed}

\item{betweenvars:}{a vector containing names of columns that are between-subjects variables}

\item{withinvars:}{a vector containing names of columns that are within-subjects variables}

\item{idvar:}{the name of a column that identifies each subject (or matched subjects)}

\item{na.rm:}{a boolean that indicates whether to ignore NA's}

\item{conf.interval:}{the percent range of the confidence interval (default is 95%)}
}
\description{
It will still work if there are no within-S variables. Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%). If there are within-subject variables, calculate adjusted values using method from Morey (2008).
}

